Separate into 2 source files
