+++ tests/Bitcoin/TWBitcoinSigningTests.cpp
@@ -9,7 +9,7 @@
 #include "Bitcoin/Script.h"
 #include "Bitcoin/Transaction.h"
 #include "Bitcoin/TransactionBuilder.h"
-#include "Bitcoin/SignatureBuilder.h"
+#include "Bitcoin/TransactionSigner.h"
 #include "Bitcoin/SigHashType.h"
 #include "Base58.h"
 #include "Hash.h"
